+++ elements-sk/error-toast-sk/index.js
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import './error-toast-sk.js'
+export { errorMessage} from './error-toast-sk.js'
