+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "elements-sk-local",
-  "version": "3.0.0",
+  "version": "2.2.0",
   "private": true,
   "description": "A set of light-weight custom elements with a uniform style.",
   "main": "index.js",
