+++ pages/index.html
@@ -20,6 +20,7 @@
 
       select-sk, multi-select-sk {
         border: 1px solid grey;
+        margin: 3px;
       }
 
     </style>
@@ -94,7 +95,7 @@
   </radiogroup>
 
   <h2 id="Toast">Toast</h2>
-  <button id=toaster>Make toast.</button>
+  <button id=toaster>Make toast</button>
   <toast-sk id=toast>
     This will disappear in 5s (the default), unless you click the Close
     button.
@@ -142,7 +143,7 @@
     <div>any</div>
     <div selected>one</div>
   </select-sk>
-  <br>
+
   <multi-select-sk>
     <div>select</div>
     <div selected>any</div>
